move a11y props
