+++ packages/lab/src/form-field-context/FormFieldContextNext.ts
@@ -1,6 +1,19 @@
 import { createContext } from "@salt-ds/core";
-import { a11yValueAriaProps } from "../form-field-next";
 
+export interface A11yValueProps {
+  /**
+   * id for FormFieldHelperText
+   */
+  helperTextId?: string;
+  /**
+   * id for FormFieldLabel
+   */
+  labelId?: string;
+}
+export interface a11yValueAriaProps {
+  "aria-labelledby": A11yValueProps["labelId"];
+  "aria-describedby": A11yValueProps["helperTextId"] | undefined;
+}
 export interface FormFieldContextNextValue {
   a11yProps: a11yValueAriaProps;
   disabled: boolean;
